Remove Exception/FilePath/AbstractFilePathException
