+++ src/Exception/FilePath/NotExistsException.php
@@ -4,10 +4,19 @@ declare(strict_types=1);
 
 namespace App\Exception\FilePath;
 
-class NotExistsException extends AbstractFilePathException
+use App\Exception\File\PathExceptionInterface;
+
+class NotExistsException extends \Exception implements PathExceptionInterface
 {
-    public function __construct(string $path, ?\Throwable $previous = null)
+    public function __construct(
+        private string $path,
+        ?\Throwable $previous = null
+    ) {
+        parent::__construct(sprintf('Path "%s" does not exist"', $path), 0, $previous);
+    }
+
+    public function getPath(): string
     {
-        parent::__construct($path, sprintf('Path "%s" does not exist"', $path), $previous);
+        return $this->path;
     }
 }
