+++ src/Exception/FilePath/OutOfScopeException.php
@@ -4,16 +4,20 @@ declare(strict_types=1);
 
 namespace App\Exception\FilePath;
 
-class OutOfScopeException extends AbstractFilePathException
+use App\Exception\File\PathExceptionInterface;
+
+class OutOfScopeException extends \Exception implements PathExceptionInterface
 {
     public function __construct(
-        string $path,
+        private string $path,
         private string $basePath
     ) {
-        parent::__construct(
-            $path,
-            sprintf('Path "%s" outside the scope of base path "%s"', $path, $basePath)
-        );
+        parent::__construct(sprintf('Path "%s" outside the scope of base path "%s"', $path, $basePath));
+    }
+
+    public function getPath(): string
+    {
+        return $this->path;
     }
 
     public function getBasePath(): string
