+++ src/Exception/FilePath/RemoveException.php
@@ -5,15 +5,21 @@ declare(strict_types=1);
 namespace App\Exception\FilePath;
 
 use App\Exception\File\MutationExceptionInterface;
+use App\Exception\File\PathExceptionInterface;
 use Symfony\Component\Filesystem\Exception\IOExceptionInterface;
 
-class RemoveException extends AbstractFilePathException implements MutationExceptionInterface
+class RemoveException extends \Exception implements MutationExceptionInterface, PathExceptionInterface
 {
     public function __construct(
-        string $path,
+        private string $path,
         private IOExceptionInterface $IOException
     ) {
-        parent::__construct($path, sprintf('Unable to remove "%s"', $path), $IOException);
+        parent::__construct(sprintf('Unable to remove "%s"', $path), 0, $IOException);
+    }
+
+    public function getPath(): string
+    {
+        return $this->path;
     }
 
     public function getIOException(): IOExceptionInterface
